fix(ir): treat if-stmt as expression in function bodies
Remove if-stmt from the statement keyword list to allow it to be used
as an expression when it appears as the sole body element of a function.

This enables recursive functions with if-stmt bodies to properly return
values from both branches instead of falling through to implicit return.

Fixes test_fn_recursive_42 which expects if-stmt to evaluate to a value.